Remove pagination, rework tests
