+++ __tests__/index.js
@@ -3,75 +3,7 @@ const payloads = require('./fixtures/payloads')
 const app = require('..')
 const fs = require('fs')
 const path = require('path')
-
-function gimmeRobot (config = 'basic.yml', issues = [{ data: {items: [{ title: 'An issue that exists', state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "An issue that exists","file": "index.js"}} -->` }], total_count: 1} }]) {
-  const cfg = config ? fs.readFileSync(path.join(__dirname, 'fixtures', 'configs', config), 'utf8') : config
-  let robot
-  let github
-  const content = (str) => Promise.resolve({ data: { content: Buffer.from(str) } })
-
-  const logger = {
-    trace: jest.fn(),
-    debug: jest.fn(),
-    info: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-    fatal: jest.fn()
-  }
-  robot = createRobot({ logger })
-  app(robot)
-
-  github = {
-    issues: {
-      getForRepo: jest.fn().mockReturnValue(Promise.resolve(issues)),
-      create: jest.fn(),
-      createLabel: jest.fn(),
-      edit: jest.fn(),
-      createComment: jest.fn()
-    },
-    search: {
-      issues: jest.fn().mockReturnValue(Promise.resolve([{data: {total_count: issues[0].length, items: issues[0].data}}]))
-    },
-    gitdata: {
-      getTree: jest.fn().mockReturnValue(Promise.resolve({
-        data: {
-          tree: [
-            { path: 'index.js', sha: 'sha' },
-            { path: 'more.js', sha: 'sha' },
-            { path: 'another.js', sha: 'sha' },
-            { path: 'many.js', sha: 'sha' },
-            { path: 'caseinsensitive.js', sha: 'sha' }
-          ]
-        }
-      })),
-      getBlob: jest.fn((obj) => ({
-        data: {
-          content: fs.readFileSync(path.join(__dirname, 'fixtures', 'files', obj.path), 'base64')
-        }
-      }))
-    },
-    paginate: jest.fn().mockReturnValue(Promise.resolve(issues)),
-    repos: {
-      // Response for getting content from '.github/todo.yml'
-      getContent: jest.fn((obj) => {
-        if (obj.path.includes('config.yml')) {
-          if (config === false) {
-            throw { code: 404 } // eslint-disable-line
-          }
-          return content(cfg)
-        } else {
-          return content(fs.readFileSync(path.join(__dirname, 'fixtures', 'files', obj.path), 'utf8'))
-        }
-      })
-    },
-    pullRequests: {
-      getAll: jest.fn().mockReturnValue(Promise.resolve({ data: [{ head: { ref: 'master' }, number: 10 }] }))
-    }
-  }
-  // Passes the mocked out GitHub API into out robot instance
-  robot.auth = () => Promise.resolve(github)
-  return { robot, github }
-}
+const {gimmeRobot} = require('./helpers')
 
 describe('todo', () => {
   it('requests issues for the repo', async () => {
@@ -174,36 +106,46 @@ describe('todo', () => {
     expect(github.issues.create).toHaveBeenCalledTimes(0)
   })
 
+  it('does not create a duplicate issue after creating one', async () => {
+    const {robot, github} = gimmeRobot('basic.yml', {data: {total_count: 0, items: []}})
+    await robot.receive(payloads.basic)
+    expect(github.issues.create).toHaveBeenCalledTimes(1)
+
+    const issue = {data: {total_count: 1, items: [{ title: 'Jason!', state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "Jason!","file": "index.js"}} -->` }]}}
+    github.search.issues.mockReturnValue(Promise.resolve(issue))
+
+    await robot.receive(payloads.basic)
+    expect(github.issues.create).toHaveBeenCalledTimes(1)
+  })
+
   it('works without a config present', async () => {
     const {robot, github} = gimmeRobot(false)
     await robot.receive(payloads.basic)
     expect(github.issues.create).toHaveBeenCalledTimes(1)
   })
 
-  it('creates 31 issues', async () => {
+  it('creates 33 issues', async () => {
     const {robot, github} = gimmeRobot()
     await robot.receive(payloads.many)
     expect(github.issues.create).toHaveBeenCalledTimes(33)
   })
 
   it('paginates when there are over 30 issues', async () => {
-    const issuesPageOne = Array.apply(null, Array(30)).map((v, i) => ({ title: `I do not exist ${i}`, state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "I do not exist ${i}"}} -->` }))
-    const issuesPageTwo = Array.apply(null, Array(3)).map((v, i) => ({ title: `I do not exist ${i + 30}`, state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "I do not exist ${i + 30}"}} -->` }))
-    const {robot, github} = gimmeRobot('basic.yml', [{data: {items: issuesPageOne, total_count: 30}}, {data: {items: issuesPageTwo, total_count: 3}}])
+    const issuesPageOne = Array.apply(null, Array(33)).map((v, i) => ({ title: `I do not exist ${i}`, state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "I do not exist ${i}"}} -->` }))
+    const {robot, github} = gimmeRobot('basic.yml', {data: {items: issuesPageOne, total_count: 33}})
     await robot.receive(payloads.many)
     expect(github.issues.create).toHaveBeenCalledTimes(33)
   })
 
   it('paginates when there are over 30 issues and does not make them', async () => {
-    const issuesPageOne = Array.apply(null, Array(30)).map((v, i) => ({ title: `I exist ${i}`, state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "I exist ${i}","file": "many.js"}} -->` }))
-    const issuesPageTwo = Array.apply(null, Array(2)).map((v, i) => ({ title: `I exist ${i + 30}`, state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "I exist ${i + 30}","file": "many.js"}} -->` }))
-    const {robot, github} = gimmeRobot('basic.yml', [{data: {items: issuesPageOne, total_count: 30}}, {data: {items: issuesPageTwo, total_count: 2}}])
+    const issuesPageOne = Array.apply(null, Array(32)).map((v, i) => ({ title: `I exist ${i}`, state: 'open', body: `\n\n<!-- probot = {"10000":{"title": "I exist ${i}","file": "many.js"}} -->` }))
+    const {robot, github} = gimmeRobot('basic.yml', {data: {items: issuesPageOne, total_count: 32}})
     await robot.receive(payloads.many)
     expect(github.issues.create).toHaveBeenCalledTimes(1)
   })
 
   it('works with issues with empty bodies', async () => {
-    const {robot, github} = gimmeRobot('basic.yml', [{data: { items: [{ title: 'Hey', state: 'open' }], total_count: 1 }}])
+    const {robot, github} = gimmeRobot('basic.yml', {data: { items: [{ title: 'Hey', state: 'open' }], total_count: 1 }})
     await robot.receive(payloads.basic)
     expect(github.issues.create).toHaveBeenCalledTimes(1)
   })
@@ -245,7 +187,7 @@ describe('todo', () => {
   })
 
   it('reopens a closed issue', async () => {
-    const issues = [{data: {
+    const issues = {data: {
       items: [{
         title: 'An issue that exists',
         state: 'open',
@@ -256,7 +198,7 @@ describe('todo', () => {
         body: `\n\n<!-- probot = {"10000":{"title": "Jason!","file": "index.js"}} -->`
       }],
       total_count: 2
-    }}]
+    }}
     const {robot, github} = gimmeRobot('basic.yml', issues)
     await robot.receive(payloads.basic)
     expect(github.issues.edit).toHaveBeenCalledTimes(1)
@@ -265,10 +207,10 @@ describe('todo', () => {
   })
 
   it('skips the search if if there are no issues', async () => {
-    const issues = [{data: {
+    const issues = {data: {
       items: [],
       total_count: 0
-    }}]
+    }}
     const {robot, github} = gimmeRobot('basic.yml', issues)
     await robot.receive(payloads.basic)
     expect(github.issues.createComment).toHaveBeenCalledTimes(0)
@@ -276,7 +218,7 @@ describe('todo', () => {
   })
 
   it('respects the reopenClosed config', async () => {
-    const issues = [{data: {
+    const issues = {data: {
       items: [{
         title: 'An issue that exists',
         state: 'open',
@@ -287,7 +229,7 @@ describe('todo', () => {
         body: `\n\n<!-- probot = {"10000":{"title": "Jason!","file": "index.js"}} -->`
       }],
       total_count: 2
-    }}]
+    }}
     const {robot, github} = gimmeRobot('reopenClosedFalse.yml', issues)
     await robot.receive(payloads.basic)
     expect(github.issues.edit).toHaveBeenCalledTimes(0)
