--- __tests__/shimRAF.js
@@ -1,3 +1,0 @@
-global.requestAnimationFrame = (callback) => {
-  setTimeout(callback, 0)
-}
