+++ package.json
@@ -21,10 +21,7 @@
   "jest": {
     "modulePathIgnorePatterns": [
       "<rootDir>/__tests__/fixtures/",
-      "<rootDir>/__tests__/shimRAF.js"
-    ],
-    "setupFiles": [
-      "<rootDir>/__tests__/shimRAF.js"
+      "<rootDir>/__tests__/helpers.js"
     ],
     "coverageDirectory": "./coverage/",
     "coveragePathIgnorePatterns": [
